requirements.txt
# Web Scraping (API-based)
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# API
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
